# docs/swagger.yaml
definitions:
...
  dto.HistoryReq:
    properties:
      id:
        type: integer
      id_trx:
        type: integer
      note:
        type: string
      status:
        type: string
    type: object
...
  models.Cinema:
    properties:
      id:
        type: integer
      logo:
        type: string
      name:
        type: string
    type: object
...
  utils.Response:
    properties:
      error:
        type: string
      message:
        type: string
      results: {}
      success:
        type: boolean
      token:
        type: string
    type: object
info:
  contact: {}
  description: This is a backend service for tickitz web app
  title: Movxtar API Documentation
  version: "1.0"
paths:
...
  /cinema:
    get:
      description: Retrieve all Cinema
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                results:
                  items:
                    $ref: '#/definitions/models.Cinema'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get  Cinema
      tags:
      - Cinema
    post:
      consumes:
      - application/json
      description: Add a new cinema to the database
      parameters:
      - description: Cinema Data
        in: body
        name: cinema
        required: true
        schema:
          $ref: '#/definitions/models.Cinema'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Token: []
      summary: Create a new cinema
      tags:
      - Cinema
  /cinema/{id}:
    delete:
      description: Delete a cinema from database by ID
      parameters:
      - description: Cinema ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Token: []
      summary: Delete a cinema
      tags:
      - Cinema
    get:
      description: Retrieve a specific cinema using its ID
      parameters:
      - description: Cinema ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                results:
                  $ref: '#/definitions/models.Cinema'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get cinema by ID
      tags:
      - Cinema
    put:
      consumes:
      - application/json
      description: Update cinema details by ID
      parameters:
      - description: Cinema ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Cinema Data
        in: body
        name: cinema
        required: true
        schema:
          $ref: '#/definitions/models.Cinema'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Token: []
      summary: Update an existing cinema
      tags:
      - Cinema
  /history:
    get:
      description: Get list of all transaction histories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Token: []
      summary: Get all history transactions
      tags:
      - History
  /history/{id}:
    patch:
      consumes:
      - application/json
      description: Update status and note of a history transaction by ID
      parameters:
      - description: History ID
        in: path
        name: id
        required: true
        type: integer
      - description: History Update Request
        in: body
        name: history
        required: true
        schema:
          $ref: '#/definitions/dto.HistoryReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Token: []
      summary: Update a history transaction status
      tags:
      - History
  /history/user:
    get:
      description: Retrieve transaction history by authenticated user ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                results:
                  items:
                    $ref: '#/definitions/dto.HistoryReq'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - Token: []
      summary: Get History by User ID
      tags:
      - History
...
securityDefinitions:
  Token:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"